AI_testowanie/requirements.txt
requests
pydantic
fastapi[standard]
uvicorn
python-dotenv
pyinstaller
llama-cpp-python
pyinstaller